requirements.txt
requests==2.28.0
python-dotenv==0.20.0
retry==0.9.2
pytest==7.1.2
